Refactor IWorkspaceRepository into multiple specific interfaces
This commit addresses the 'God Object' anti-pattern present in `IWorkspaceRepository` and its concrete implementation `WorkspaceRepository` by applying the Single Responsibility Principle (SRP) and Interface Segregation Principle (ISP).

The monolithic repository has been split into five domain-specific repositories:
- IWorkspaceManagementRepository
- IWorkspaceMemberRepository
- IPropertyRepository
- ISavedPropertyRepository
- IActivityLogRepository

All dependent application services (`WorkspaceService`, `WorkspaceMemberService`, `WorkspacePropertyService`), extension methods, notification event handlers, and unit tests have been updated to inject and use these specific repositories. Dependency injection registrations have also been updated to map the new interfaces correctly.
