+++ backend/Valora.Application/Services/WorkspaceMemberService.cs
@@ -7,13 +7,22 @@ namespace Valora.Application.Services;
 
 public class WorkspaceMemberService : IWorkspaceMemberService
 {
-    private readonly IWorkspaceRepository _repository;
+    private readonly IWorkspaceManagementRepository _managementRepository;
+    private readonly IWorkspaceMemberRepository _memberRepository;
+    private readonly IActivityLogRepository _activityLogRepository;
     private readonly IIdentityService _identityService;
     private readonly IEventDispatcher _eventDispatcher;
 
-    public WorkspaceMemberService(IWorkspaceRepository repository, IIdentityService identityService, IEventDispatcher eventDispatcher)
+    public WorkspaceMemberService(
+        IWorkspaceManagementRepository managementRepository,
+        IWorkspaceMemberRepository memberRepository,
+        IActivityLogRepository activityLogRepository,
+        IIdentityService identityService,
+        IEventDispatcher eventDispatcher)
     {
-        _repository = repository;
+        _managementRepository = managementRepository;
+        _memberRepository = memberRepository;
+        _activityLogRepository = activityLogRepository;
         _identityService = identityService;
         _eventDispatcher = eventDispatcher;
     }
@@ -22,7 +31,7 @@ public class WorkspaceMemberService : IWorkspaceMemberService
     {
         await ValidateMemberAccess(userId, workspaceId, ct);
 
-        var members = await _repository.GetMembersAsync(workspaceId, ct);
+        var members = await _memberRepository.GetMembersAsync(workspaceId, ct);
 
         return members.Select(m => new WorkspaceMemberDto(
             m.Id,
@@ -39,7 +48,7 @@ public class WorkspaceMemberService : IWorkspaceMemberService
         var role = await GetUserRole(userId, workspaceId, ct);
         if (role != WorkspaceRole.Owner) throw new ForbiddenAccessException();
 
-        var existingMember = await _repository.GetMemberByEmailAsync(workspaceId, dto.Email, ct);
+        var existingMember = await _memberRepository.GetMemberByEmailAsync(workspaceId, dto.Email, ct);
         if (existingMember != null) return;
 
         var invitedUser = await _identityService.GetUserByEmailAsync(dto.Email);
@@ -53,11 +62,11 @@ public class WorkspaceMemberService : IWorkspaceMemberService
             JoinedAt = invitedUser != null ? DateTime.UtcNow : (DateTime?)null
         };
 
-        await _repository.AddMemberAsync(member, ct);
+        await _memberRepository.AddMemberAsync(member, ct);
         await LogActivityAsync(workspaceId, userId, ActivityLogType.MemberInvited, $"Invited a new member as {dto.Role}", ct);
-        await _repository.SaveChangesAsync(ct);
+        await _memberRepository.SaveChangesAsync(ct);
 
-        var workspace = await _repository.GetByIdAsync(workspaceId, ct);
+        var workspace = await _managementRepository.GetByIdAsync(workspaceId, ct);
         if (workspace != null)
         {
             await _eventDispatcher.DispatchAsync(new Valora.Application.Common.Events.WorkspaceInviteAcceptedEvent(workspaceId, workspace.Name, userId, member.UserId ?? "", dto.Email), ct);
@@ -69,26 +78,26 @@ public class WorkspaceMemberService : IWorkspaceMemberService
         var currentUserRole = await GetUserRole(userId, workspaceId, ct);
         if (currentUserRole != WorkspaceRole.Owner) throw new ForbiddenAccessException();
 
-        var member = await _repository.GetMemberAsync(memberId, ct);
+        var member = await _memberRepository.GetMemberAsync(memberId, ct);
         if (member == null || member.WorkspaceId != workspaceId) throw new NotFoundException(nameof(WorkspaceMember), memberId);
 
         if (member.UserId == userId) throw new InvalidOperationException("Cannot remove yourself.");
 
-        await _repository.RemoveMemberAsync(member, ct);
+        await _memberRepository.RemoveMemberAsync(member, ct);
         await LogActivityAsync(workspaceId, userId, ActivityLogType.MemberRemoved, "Removed a member", ct);
-        await _repository.SaveChangesAsync(ct);
+        await _memberRepository.SaveChangesAsync(ct);
     }
 
     // Helpers
     private async Task ValidateMemberAccess(string userId, Guid workspaceId, CancellationToken ct)
     {
-        var isMember = await _repository.IsMemberAsync(workspaceId, userId, ct);
+        var isMember = await _memberRepository.IsMemberAsync(workspaceId, userId, ct);
         if (!isMember) throw new ForbiddenAccessException();
     }
 
     private async Task<WorkspaceRole> GetUserRole(string userId, Guid workspaceId, CancellationToken ct)
     {
-        return await _repository.GetUserRoleAsync(workspaceId, userId, ct);
+        return await _memberRepository.GetUserRoleAsync(workspaceId, userId, ct);
     }
 
     private async Task LogActivityAsync(Guid? workspaceId, string actorId, ActivityLogType type, string summary, CancellationToken ct)
@@ -101,6 +110,6 @@ public class WorkspaceMemberService : IWorkspaceMemberService
             Summary = summary,
             CreatedAt = DateTime.UtcNow
         };
-        await _repository.LogActivityAsync(log, ct);
+        await _activityLogRepository.LogActivityAsync(log, ct);
     }
 }
